Update to mixed v1/v2 API
Twitter recently suspended accounts still using the v1 API (including
@MOF_papers). So the migration is underway, and the current code is
designed to work with the "Free" Twitter plan. It uses the v2 API to
post tweets.

Sadly, the v2 API does not allow to upload media yet. So the v1 API is
retained and still used to upload media, which are attached to tweets
sent with the v2 API. Reference: https://stackoverflow.com/a/76509550